requirements.txt
fastapi
uvicorn[standard]
python-dotenv
llama-cpp-python
chromadb
websockets